Revert "fixed enhanced exit option"
This reverts commit b49af48835a2f9e78c21eb0dcb79af89a8b5247d.
