+++ PowerUp/app/src/main/java/powerup/systers/com/AboutActivity.java
@@ -1,5 +1,6 @@
 package powerup.systers.com;
 
+import android.app.Activity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
