+++ PowerUp/app/src/main/java/powerup/systers/com/AvatarRoomActivity.java
@@ -5,6 +5,7 @@
 
 package powerup.systers.com;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
