+++ PowerUp/app/src/main/java/powerup/systers/com/GameActivity.java
@@ -6,6 +6,7 @@
 package powerup.systers.com;
 
 import android.annotation.SuppressLint;
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
