+++ PowerUp/app/src/main/java/powerup/systers/com/MapActivity.java
@@ -4,6 +4,7 @@
 
 package powerup.systers.com;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
